refactor(daemon): extract the memory L0 type contract into memory/types.ts
The pure contract at the top of `memory/provider.ts` — MemoryScope,
RecallRequest, the read/write results, TurnRecord, the admin surfaces, the
MemoryProvider port and MemoryProviderUnavailableError — now lives in
`memory/types.ts`. Implementations and factories stay in `provider.ts`, which
re-exports every moved name so its ~17 external importers are untouched.

That kills the two type-only back-edges into `provider.ts`:
`memory/runtime/native.ts` and `memory/recall.ts` now import from `./types.js`
directly, so neither leaf pulls the implementation module for a type.

Verified: daemon typecheck clean; memory-provider-dispatch, external-memory-provider,
memory-recall and memory tests pass (79); prettier + eslint clean on touched files.
